Keep provider-turn lane execution out of the coordinator root
The coordinator still carried a full copy of the provider-turn lane execution shell even after the surrounding runtime, reply, and apply seams had been extracted. This move makes provider_turn_lane.rs the authoritative home for lane execution and malformed-followup followup detection so the coordinator root keeps shrinking toward orchestration.

Constraint: Preserve provider-turn lane execution behavior, including malformed-followup and discovery recovery paths
Rejected: Leave the lane shell inline until safe-lane extraction also happens | that would keep a fully extracted seam half-owned by the coordinator root
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep provider-turn lane execution in provider_turn_lane.rs and avoid reintroducing a shadow implementation in turn_coordinator.rs
Tested: ./scripts/cargo-local-toolchain.sh check -p loong-app; ./scripts/cargo-local-toolchain.sh test -p loong-app provider_turn_lane_plan -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app build_provider_turn_tool_terminal_events -- --test-threads=1; ./scripts/cargo-local-toolchain.sh fmt --all -- --check; ./scripts/cargo-local-toolchain.sh clippy -p loong-app --all-targets -- -D warnings
Not-tested: full workspace test suite; deeper safe-lane execution matrix in this slice
